setting up for deployment and adding alerts

+++ frontend/src/store/listSlice.js
@@ -1,26 +1,28 @@
 import { createSlice } from "@reduxjs/toolkit";
 
 const loadListFromLocalStorage = () => {
-  const savedList = localStorage.getItem('list');
+  const savedList = localStorage.getItem("list");
   return savedList ? JSON.parse(savedList) : [];
 };
 
 const listSlice = createSlice({
-  name: 'list',
+  name: "list",
   initialState: loadListFromLocalStorage(),
-  reducers:{
-    add(state, action){
-      state.push(action.payload);
-      localStorage.setItem('list', JSON.stringify(state)); 
+  reducers: {
+    add(state, action) {
+      const exists = state.some((university) => university.id === action.payload.id);
+      if (!exists) {
+        state.push(action.payload);
+        localStorage.setItem("list", JSON.stringify(state));
+      }
     },
     remove(state, action) {
       const updatedList = state.filter((product) => product.id !== action.payload.id);
-      localStorage.setItem("list", JSON.stringify(updatedList)); 
-      return updatedList; 
-    }
-    
-  }
-})
+      localStorage.setItem("list", JSON.stringify(updatedList));
+      return updatedList;
+    },
+  },
+});
 
 export const { add, remove } = listSlice.actions;
 export default listSlice.reducer;
